page-manager renamed
