+++ e2e/tests/home/home-page.spec.ts
@@ -1,6 +1,6 @@
 import HomePage from 'app/page/home-page';
 import { Page } from 'puppeteer';
-import { withPage } from 'libs/test-page-manager';
+import { withPage } from 'libs/page-manager';
 import { signIn } from 'utils/test-utils';
 import { config } from 'resources/workbench-config';
